// src/components/models/StudentDetailModel.ts
export interface StudentDetailModel {
    id: number;
    studentId: string;
    firstName: string;
    middleName: string;
    lastName: string;
    course: string;
    yearLevel: string;
    section: string;
    schoolEmail: string;
    personalEmail: string;
    portfolioURL: string;
    profilePicture: string;    
  }